+++ src/Core/Application/Models/DTOs/Assets/UserAssetForOperationDto.cs
@@ -7,7 +7,7 @@ namespace Application.Models.DTOs.Assets
 
         }
 
-        public UserAssetForOperationDto(int currencyId, string title, string? subTitle, int count, decimal purchasePrice, decimal salePrice)
+        public UserAssetForOperationDto(int currencyId, string title, string? subTitle, decimal count, decimal purchasePrice, decimal salePrice)
         {
             CurrencyId = currencyId;
             Title = title;
@@ -20,7 +20,7 @@ namespace Application.Models.DTOs.Assets
         public int CurrencyId { get; init; }
         public string Title { get; init; } = null!;
         public string? SubTitle { get; init; }
-        public int Count { get; init; }
+        public decimal Count { get; init; }
         public decimal PurchasePrice { get; init; }
         public decimal SalePrice { get; init; }
     }
